localsearch works now
git-svn-id: https://svn.ipd.kit.edu/repos/karcher/karcher/markus/Tuner@2332 1cde74c2-b0a7-11dd-9e77-cbfca2305f68

+++ Server/LocalSearch.h
@@ -4,13 +4,13 @@
 
 class LocalSearch {
 	public:
-		LocalSearch(McHandler* handler, double threshold, int retryCount);
+		LocalSearch(McHandler* handler, int threshold, int retryCount);
 		~LocalSearch();
 		int doLocalSearch();
 
 	private:
 		McHandler* mcHandler;
-		double threshold;
+		int threshold;
 		int retryCount;
 		opt_mc_t* bestMc;
 		opt_mc_t* curMc;
