+++ utils.h
@@ -11,4 +11,4 @@ using namespace std;
 void errorExit(string msg); 
 void diff(timespec* start, timespec* end, timespec* diff);
 bool isTimespecLower(timespec* first, timespec* second);
-double getRelativePerformance(timespec* first, timespec* second);
+int getRelativePerformance(timespec* first, timespec* second);
